+++ plugins/newspack-network/CHANGELOG.md
@@ -1,3 +1,24 @@
+## newspack-network [2.22.6](https://github.com/Automattic/newspack-workspace/compare/newspack-network@2.22.5...newspack-network@2.22.6) (2026-09-02)
+
+
+### Bug Fixes
+
+* **network:** harden remote media handling in distribution ([#885](https://github.com/Automattic/newspack-workspace/issues/885)) ([f450a66](https://github.com/Automattic/newspack-workspace/commit/f450a663c6c876d89ca26b549c9c23aac17cab7a))
+
+## newspack-network [2.22.5](https://github.com/Automattic/newspack-workspace/compare/newspack-network@2.22.4...newspack-network@2.22.5) (2026-08-31)
+
+
+### Dependencies
+
+* **newspack-icons:** upgraded to 1.1.2
+
+## newspack-network [2.22.4](https://github.com/Automattic/newspack-workspace/compare/newspack-network@2.22.3...newspack-network@2.22.4) (2026-08-26)
+
+
+### Bug Fixes
+
+* **network:** process each hub webhook delivery at most once ([#940](https://github.com/Automattic/newspack-workspace/issues/940)) ([6976c7b](https://github.com/Automattic/newspack-workspace/commit/6976c7bc35117b0dc859fa8903e13c3071cd3731))
+
 ## newspack-network [2.22.3](https://github.com/Automattic/newspack-workspace/compare/newspack-network@2.22.2...newspack-network@2.22.3) (2026-08-26)